\name{SimpleStatTools-package}
\alias{SimpleStatTools-package}
\alias{SimpleStatTools}
\docType{package}
\title{
\packageTitle{SimpleStatTools}
}
\description{
\packageDescription{SimpleStatTools}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SimpleStatTools}
\packageIndices{SimpleStatTools}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{SimpleStatTools}

Maintainer: \packageMaintainer{SimpleStatTools}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
